softdesk_api/softdesk_api/urls.py
from django.contrib import admin
from django.urls import path
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView

from users.views import DataUserAPIView, RegisterView

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/users/', DataUserAPIView.as_view()),
    path('api/signup/', RegisterView.as_view(), name='signup'),
    path('api/login/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
]
