pyproject.toml
[project]
name = "sofdesk-api-dependencies"
version = "2.1.3"
description = "project dependancies"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "None"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
